add clean and object-type options to extract
